test: refactor common utils
