+++ test/factory/utils.ts
@@ -1,10 +1,24 @@
-import { constants } from 'ethers';
+import { constants, Wallet } from 'ethers';
+import seedrandom from 'seedrandom';
+import { getSuffixData } from '../smartwallet/utils';
+import {
+  getLocalEip712DeploySignature,
+  TypedDeployRequestData,
+} from '../utils/EIP712Utils';
+
 import {
   EnvelopingTypes,
   IForwarder,
 } from 'typechain-types/contracts/RelayHub';
 
-export function createDeployRequest(
+const random = seedrandom('rif');
+const minIndex = 0;
+const maxIndex = 1000000000;
+
+const randomNumber = (min = minIndex, max = maxIndex) =>
+  Math.floor(random() * (max - min + 1) + min);
+
+function createDeployRequest(
   request: Partial<IForwarder.DeployRequestStruct>,
   relayData?: Partial<EnvelopingTypes.RelayDataStruct>
 ): EnvelopingTypes.DeployRequestStruct {
@@ -42,3 +56,29 @@ export function createDeployRequest(
     },
   };
 }
+
+const signDeployRequest = (
+  signer: Wallet,
+  deployRequest: EnvelopingTypes.DeployRequestStruct,
+  verifier: string,
+  chainId: number
+) => {
+  const typedDeployData = new TypedDeployRequestData(
+    chainId,
+    verifier,
+    deployRequest
+  );
+
+  const suffixData = getSuffixData(typedDeployData);
+
+  const privateKey = Buffer.from(signer.privateKey.substring(2, 66), 'hex');
+
+  const signature = getLocalEip712DeploySignature(typedDeployData, privateKey);
+
+  return {
+    suffixData,
+    signature,
+  };
+};
+
+export { randomNumber, createDeployRequest, signDeployRequest };
